Delay the config parser data read
